+++ architect_py/__init__.py
@@ -1,17 +1,442 @@
-__version__ = "5.0.0"
+# ruff: noqa:I001
 
+__version__ = "5.1.0b1"
+
+from .utils.nearest_tick import TickRoundMethod
 from .async_client import AsyncClient
 from .client import Client
-from .common_types import OrderDir, TradableProduct
-from .grpc import CandleWidth, L2BookDiff, L2BookSnapshot, OrderStatus
+from .common_types import OrderDir, TradableProduct, TimeInForce, Venue
+from .grpc.models.definitions import (
+    AccountIdOrName,
+    AccountPosition,
+    AccountStatistics,
+    AlgoOrderStatus,
+    CancelStatus,
+    CandleWidth,
+    CptyLogoutRequest,
+    DateTimeOrUtc,
+    Deposit,
+    HealthMetric,
+    HealthStatus,
+    L2BookDiff,
+    OrderId,
+    OrderOut,
+    OrderRejectReason,
+    OrderSource,
+    OrderStale,
+    OrderStatus,
+    OrderType,
+    ProductCatalogInfo,
+    RqdAccountStatistics,
+    SortTickersBy,
+    Statement,
+    TraderIdOrEmail,
+    UserId,
+    Withdrawal,
+    AccountPermissions,
+    AliasKind,
+    DerivativeKind,
+    FillKind,
+    Unit,
+    MinOrderQuantityUnit,
+    OptionsExerciseType,
+    PriceDisplayFormat,
+    Fiat,
+    Commodity,
+    Crypto,
+    Equity,
+    Index,
+    Future,
+    Perpetual,
+    Unknown,
+    PutOrCall,
+    SnapshotOrUpdateForStringAndProductCatalogInfo1,
+    SnapshotOrUpdateForStringAndProductCatalogInfo2,
+    SnapshotOrUpdateForStringAndString1,
+    SnapshotOrUpdateForStringAndString2,
+    SimpleDecimal,
+    Varying1,
+    Varying,
+    AccountName,
+    OptionLike,
+    EventContractSeriesInstance2,
+    OptionsSeriesInstance,
+    SpreadLeg,
+    Outcome,
+    AberrantFill,
+    CancelReject,
+    CptyLoginRequest,
+    ExecutionInfo,
+    Fill,
+    OptionsContract,
+    OptionsGreeks,
+    OptionsSeriesInfo,
+    OrderAck,
+    OrderCanceled,
+    OrderCanceling,
+    OrderReject,
+    SnapshotOrUpdateForAliasKindAndSnapshotOrUpdateForStringAndString1,
+    SnapshotOrUpdateForAliasKindAndSnapshotOrUpdateForStringAndString2,
+    SnapshotOrUpdateForStringAndOptionsSeriesInfo1,
+    SnapshotOrUpdateForStringAndOptionsSeriesInfo2,
+    SnapshotOrUpdateForStringAndSnapshotOrUpdateForStringAndProductCatalogInfo1,
+    SnapshotOrUpdateForStringAndSnapshotOrUpdateForStringAndProductCatalogInfo2,
+    Account,
+    FutureSpread,
+    Option,
+    SnapshotOrUpdateForStringAndExecutionInfo1,
+    SnapshotOrUpdateForStringAndExecutionInfo2,
+    Enumerated,
+    EventContractSeriesInstance1,
+    AccountWithPermissions,
+    SnapshotOrUpdateForStringAndSnapshotOrUpdateForStringAndExecutionInfo1,
+    SnapshotOrUpdateForStringAndSnapshotOrUpdateForStringAndExecutionInfo2,
+    EventContract,
+    ProductInfo,
+    SnapshotOrUpdateForStringAndProductInfo1,
+    SnapshotOrUpdateForStringAndProductInfo2,
+)
+from .grpc.models.Accounts.AccountsRequest import AccountsRequest
+from .grpc.models.Accounts.AccountsResponse import AccountsResponse
+from .grpc.models.Algo.AlgoOrder import AlgoOrder
+from .grpc.models.Algo.AlgoOrderRequest import AlgoOrderRequest
+from .grpc.models.Algo.AlgoOrdersRequest import AlgoOrdersRequest
+from .grpc.models.Algo.AlgoOrdersResponse import AlgoOrdersResponse
+from .grpc.models.Algo.CreateAlgoOrderRequest import CreateAlgoOrderRequest
+from .grpc.models.Algo.PauseAlgoRequest import PauseAlgoRequest
+from .grpc.models.Algo.PauseAlgoResponse import PauseAlgoResponse
+from .grpc.models.Algo.StartAlgoRequest import StartAlgoRequest
+from .grpc.models.Algo.StartAlgoResponse import StartAlgoResponse
+from .grpc.models.Algo.StopAlgoRequest import StopAlgoRequest
+from .grpc.models.Algo.StopAlgoResponse import StopAlgoResponse
+from .grpc.models.Auth.CreateJwtRequest import CreateJwtRequest
+from .grpc.models.Auth.CreateJwtResponse import CreateJwtResponse
+from .grpc.models.Boss.DepositsRequest import DepositsRequest
+from .grpc.models.Boss.DepositsResponse import DepositsResponse
+from .grpc.models.Boss.RqdAccountStatisticsRequest import RqdAccountStatisticsRequest
+from .grpc.models.Boss.RqdAccountStatisticsResponse import RqdAccountStatisticsResponse
+from .grpc.models.Boss.StatementUrlRequest import StatementUrlRequest
+from .grpc.models.Boss.StatementUrlResponse import StatementUrlResponse
+from .grpc.models.Boss.StatementsRequest import StatementsRequest
+from .grpc.models.Boss.StatementsResponse import StatementsResponse
+from .grpc.models.Boss.WithdrawalsRequest import WithdrawalsRequest
+from .grpc.models.Boss.WithdrawalsResponse import WithdrawalsResponse
+from .grpc.models.Core.ConfigRequest import ConfigRequest
+from .grpc.models.Core.ConfigResponse import ConfigResponse
+from .grpc.models.Core.RestartCptyRequest import RestartCptyRequest
+from .grpc.models.Core.RestartCptyResponse import RestartCptyResponse
+from .grpc.models.Cpty.CptyRequest import CptyRequest
+from .grpc.models.Cpty.CptyResponse import CptyResponse
+from .grpc.models.Cpty.CptyStatus import CptyStatus
+from .grpc.models.Cpty.CptyStatusRequest import CptyStatusRequest
+from .grpc.models.Cpty.CptysRequest import CptysRequest
+from .grpc.models.Cpty.CptysResponse import CptysResponse
+from .grpc.models.Folio.AccountHistoryRequest import AccountHistoryRequest
+from .grpc.models.Folio.AccountHistoryResponse import AccountHistoryResponse
+from .grpc.models.Folio.AccountSummariesRequest import AccountSummariesRequest
+from .grpc.models.Folio.AccountSummariesResponse import AccountSummariesResponse
+from .grpc.models.Folio.AccountSummary import AccountSummary
+from .grpc.models.Folio.AccountSummaryRequest import AccountSummaryRequest
+from .grpc.models.Folio.HistoricalFillsRequest import HistoricalFillsRequest
+from .grpc.models.Folio.HistoricalFillsResponse import HistoricalFillsResponse
+from .grpc.models.Folio.HistoricalOrdersRequest import HistoricalOrdersRequest
+from .grpc.models.Folio.HistoricalOrdersResponse import HistoricalOrdersResponse
+from .grpc.models.Health.HealthCheckRequest import HealthCheckRequest
+from .grpc.models.Health.HealthCheckResponse import HealthCheckResponse
+from .grpc.models.Marketdata.ArrayOfL1BookSnapshot import ArrayOfL1BookSnapshot
+from .grpc.models.Marketdata.Candle import Candle
+from .grpc.models.Marketdata.HistoricalCandlesRequest import HistoricalCandlesRequest
+from .grpc.models.Marketdata.HistoricalCandlesResponse import HistoricalCandlesResponse
+from .grpc.models.Marketdata.L1BookSnapshot import L1BookSnapshot
+from .grpc.models.Marketdata.L1BookSnapshotRequest import L1BookSnapshotRequest
+from .grpc.models.Marketdata.L1BookSnapshotsRequest import L1BookSnapshotsRequest
+from .grpc.models.Marketdata.L2BookSnapshot import L2BookSnapshot
+from .grpc.models.Marketdata.L2BookSnapshotRequest import L2BookSnapshotRequest
+from .grpc.models.Marketdata.L2BookUpdate import L2BookUpdate
+from .grpc.models.Marketdata.Liquidation import Liquidation
+from .grpc.models.Marketdata.MarketStatus import MarketStatus
+from .grpc.models.Marketdata.MarketStatusRequest import MarketStatusRequest
+from .grpc.models.Marketdata.SubscribeCandlesRequest import SubscribeCandlesRequest
+from .grpc.models.Marketdata.SubscribeCurrentCandlesRequest import (
+    SubscribeCurrentCandlesRequest,
+)
+from .grpc.models.Marketdata.SubscribeL1BookSnapshotsRequest import (
+    SubscribeL1BookSnapshotsRequest,
+)
+from .grpc.models.Marketdata.SubscribeL2BookUpdatesRequest import (
+    SubscribeL2BookUpdatesRequest,
+)
+from .grpc.models.Marketdata.SubscribeLiquidationsRequest import (
+    SubscribeLiquidationsRequest,
+)
+from .grpc.models.Marketdata.SubscribeManyCandlesRequest import (
+    SubscribeManyCandlesRequest,
+)
+from .grpc.models.Marketdata.SubscribeTickersRequest import SubscribeTickersRequest
+from .grpc.models.Marketdata.SubscribeTradesRequest import SubscribeTradesRequest
+from .grpc.models.Marketdata.Ticker import Ticker
+from .grpc.models.Marketdata.TickerRequest import TickerRequest
+from .grpc.models.Marketdata.TickerUpdate import TickerUpdate
+from .grpc.models.Marketdata.TickersRequest import TickersRequest
+from .grpc.models.Marketdata.TickersResponse import TickersResponse
+from .grpc.models.Marketdata.Trade import Trade
+from .grpc.models.Oms.Cancel import Cancel
+from .grpc.models.Oms.CancelAllOrdersRequest import CancelAllOrdersRequest
+from .grpc.models.Oms.CancelAllOrdersResponse import CancelAllOrdersResponse
+from .grpc.models.Oms.CancelOrderRequest import CancelOrderRequest
+from .grpc.models.Oms.OpenOrdersRequest import OpenOrdersRequest
+from .grpc.models.Oms.OpenOrdersResponse import OpenOrdersResponse
+from .grpc.models.Oms.Order import Order
+from .grpc.models.Oms.PendingCancelsRequest import PendingCancelsRequest
+from .grpc.models.Oms.PendingCancelsResponse import PendingCancelsResponse
+from .grpc.models.Oms.PlaceOrderRequest import PlaceOrderRequest
+from .grpc.models.OptionsMarketdata.OptionsChain import OptionsChain
+from .grpc.models.OptionsMarketdata.OptionsChainGreeks import OptionsChainGreeks
+from .grpc.models.OptionsMarketdata.OptionsChainGreeksRequest import (
+    OptionsChainGreeksRequest,
+)
+from .grpc.models.OptionsMarketdata.OptionsChainRequest import OptionsChainRequest
+from .grpc.models.OptionsMarketdata.OptionsExpirations import OptionsExpirations
+from .grpc.models.OptionsMarketdata.OptionsExpirationsRequest import (
+    OptionsExpirationsRequest,
+)
+from .grpc.models.Orderflow.Dropcopy import Dropcopy
+from .grpc.models.Orderflow.DropcopyRequest import DropcopyRequest
+from .grpc.models.Orderflow.Orderflow import Orderflow
+from .grpc.models.Orderflow.OrderflowRequest import OrderflowRequest
+from .grpc.models.Orderflow.SubscribeOrderflowRequest import SubscribeOrderflowRequest
+from .grpc.models.Symbology.DownloadProductCatalogRequest import (
+    DownloadProductCatalogRequest,
+)
+from .grpc.models.Symbology.DownloadProductCatalogResponse import (
+    DownloadProductCatalogResponse,
+)
+from .grpc.models.Symbology.ExecutionInfoRequest import ExecutionInfoRequest
+from .grpc.models.Symbology.ExecutionInfoResponse import ExecutionInfoResponse
+from .grpc.models.Symbology.PruneExpiredSymbolsRequest import PruneExpiredSymbolsRequest
+from .grpc.models.Symbology.PruneExpiredSymbolsResponse import (
+    PruneExpiredSymbolsResponse,
+)
+from .grpc.models.Symbology.SubscribeSymbology import SubscribeSymbology
+from .grpc.models.Symbology.SymbologyRequest import SymbologyRequest
+from .grpc.models.Symbology.SymbologySnapshot import SymbologySnapshot
+from .grpc.models.Symbology.SymbologyUpdate import SymbologyUpdate
+from .grpc.models.Symbology.SymbolsRequest import SymbolsRequest
+from .grpc.models.Symbology.SymbolsResponse import SymbolsResponse
+from .grpc.models.Symbology.UploadProductCatalogRequest import (
+    UploadProductCatalogRequest,
+)
+from .grpc.models.Symbology.UploadProductCatalogResponse import (
+    UploadProductCatalogResponse,
+)
+from .grpc.models.Symbology.UploadSymbologyRequest import UploadSymbologyRequest
+from .grpc.models.Symbology.UploadSymbologyResponse import UploadSymbologyResponse
 
 __all__ = [
+    "AberrantFill",
+    "Account",
+    "AccountHistoryRequest",
+    "AccountHistoryResponse",
+    "AccountIdOrName",
+    "AccountName",
+    "AccountPermissions",
+    "AccountPosition",
+    "AccountStatistics",
+    "AccountSummariesRequest",
+    "AccountSummariesResponse",
+    "AccountSummary",
+    "AccountSummaryRequest",
+    "AccountWithPermissions",
+    "AccountsRequest",
+    "AccountsResponse",
+    "AlgoOrder",
+    "AlgoOrderRequest",
+    "AlgoOrderStatus",
+    "AlgoOrdersRequest",
+    "AlgoOrdersResponse",
+    "AliasKind",
+    "ArrayOfL1BookSnapshot",
     "AsyncClient",
-    "Client",
+    "Cancel",
+    "CancelAllOrdersRequest",
+    "CancelAllOrdersResponse",
+    "CancelOrderRequest",
+    "CancelReject",
+    "CancelStatus",
+    "Candle",
     "CandleWidth",
-    "OrderDir",
-    "OrderStatus",
-    "TradableProduct",
-    "L2BookSnapshot",
+    "Client",
+    "Commodity",
+    "ConfigRequest",
+    "ConfigResponse",
+    "CptyLoginRequest",
+    "CptyLogoutRequest",
+    "CptyRequest",
+    "CptyResponse",
+    "CptyStatus",
+    "CptyStatusRequest",
+    "CptysRequest",
+    "CptysResponse",
+    "CreateAlgoOrderRequest",
+    "CreateJwtRequest",
+    "CreateJwtResponse",
+    "Crypto",
+    "DateTimeOrUtc",
+    "Deposit",
+    "DepositsRequest",
+    "DepositsResponse",
+    "DerivativeKind",
+    "DownloadProductCatalogRequest",
+    "DownloadProductCatalogResponse",
+    "Dropcopy",
+    "DropcopyRequest",
+    "Enumerated",
+    "Equity",
+    "EventContract",
+    "EventContractSeriesInstance1",
+    "EventContractSeriesInstance2",
+    "ExecutionInfo",
+    "ExecutionInfoRequest",
+    "ExecutionInfoResponse",
+    "Fiat",
+    "Fill",
+    "FillKind",
+    "Future",
+    "FutureSpread",
+    "HealthCheckRequest",
+    "HealthCheckResponse",
+    "HealthMetric",
+    "HealthStatus",
+    "HistoricalCandlesRequest",
+    "HistoricalCandlesResponse",
+    "HistoricalFillsRequest",
+    "HistoricalFillsResponse",
+    "HistoricalOrdersRequest",
+    "HistoricalOrdersResponse",
+    "Index",
+    "L1BookSnapshot",
+    "L1BookSnapshotRequest",
+    "L1BookSnapshotsRequest",
     "L2BookDiff",
+    "L2BookSnapshot",
+    "L2BookSnapshotRequest",
+    "L2BookUpdate",
+    "Liquidation",
+    "MarketStatus",
+    "MarketStatusRequest",
+    "MinOrderQuantityUnit",
+    "OpenOrdersRequest",
+    "OpenOrdersResponse",
+    "Option",
+    "OptionLike",
+    "OptionsChain",
+    "OptionsChainGreeks",
+    "OptionsChainGreeksRequest",
+    "OptionsChainRequest",
+    "OptionsContract",
+    "OptionsExerciseType",
+    "OptionsExpirations",
+    "OptionsExpirationsRequest",
+    "OptionsGreeks",
+    "OptionsSeriesInfo",
+    "OptionsSeriesInstance",
+    "Order",
+    "OrderAck",
+    "OrderCanceled",
+    "OrderCanceling",
+    "OrderDir",
+    "OrderId",
+    "OrderOut",
+    "OrderReject",
+    "OrderRejectReason",
+    "OrderSource",
+    "OrderStale",
+    "OrderStatus",
+    "OrderType",
+    "Orderflow",
+    "OrderflowRequest",
+    "Outcome",
+    "PauseAlgoRequest",
+    "PauseAlgoResponse",
+    "PendingCancelsRequest",
+    "PendingCancelsResponse",
+    "Perpetual",
+    "PlaceOrderRequest",
+    "PriceDisplayFormat",
+    "ProductCatalogInfo",
+    "ProductInfo",
+    "PruneExpiredSymbolsRequest",
+    "PruneExpiredSymbolsResponse",
+    "PutOrCall",
+    "RestartCptyRequest",
+    "RestartCptyResponse",
+    "RqdAccountStatistics",
+    "RqdAccountStatisticsRequest",
+    "RqdAccountStatisticsResponse",
+    "SimpleDecimal",
+    "SnapshotOrUpdateForAliasKindAndSnapshotOrUpdateForStringAndString1",
+    "SnapshotOrUpdateForAliasKindAndSnapshotOrUpdateForStringAndString2",
+    "SnapshotOrUpdateForStringAndExecutionInfo1",
+    "SnapshotOrUpdateForStringAndExecutionInfo2",
+    "SnapshotOrUpdateForStringAndOptionsSeriesInfo1",
+    "SnapshotOrUpdateForStringAndOptionsSeriesInfo2",
+    "SnapshotOrUpdateForStringAndProductCatalogInfo1",
+    "SnapshotOrUpdateForStringAndProductCatalogInfo2",
+    "SnapshotOrUpdateForStringAndProductInfo1",
+    "SnapshotOrUpdateForStringAndProductInfo2",
+    "SnapshotOrUpdateForStringAndSnapshotOrUpdateForStringAndExecutionInfo1",
+    "SnapshotOrUpdateForStringAndSnapshotOrUpdateForStringAndExecutionInfo2",
+    "SnapshotOrUpdateForStringAndSnapshotOrUpdateForStringAndProductCatalogInfo1",
+    "SnapshotOrUpdateForStringAndSnapshotOrUpdateForStringAndProductCatalogInfo2",
+    "SnapshotOrUpdateForStringAndString1",
+    "SnapshotOrUpdateForStringAndString2",
+    "SortTickersBy",
+    "SpreadLeg",
+    "StartAlgoRequest",
+    "StartAlgoResponse",
+    "Statement",
+    "StatementUrlRequest",
+    "StatementUrlResponse",
+    "StatementsRequest",
+    "StatementsResponse",
+    "StopAlgoRequest",
+    "StopAlgoResponse",
+    "SubscribeCandlesRequest",
+    "SubscribeCurrentCandlesRequest",
+    "SubscribeL1BookSnapshotsRequest",
+    "SubscribeL2BookUpdatesRequest",
+    "SubscribeLiquidationsRequest",
+    "SubscribeManyCandlesRequest",
+    "SubscribeOrderflowRequest",
+    "SubscribeSymbology",
+    "SubscribeTickersRequest",
+    "SubscribeTradesRequest",
+    "SymbologyRequest",
+    "SymbologySnapshot",
+    "SymbologyUpdate",
+    "SymbolsRequest",
+    "SymbolsResponse",
+    "TickRoundMethod",
+    "Ticker",
+    "TickerRequest",
+    "TickerUpdate",
+    "TickersRequest",
+    "TickersResponse",
+    "TimeInForce",
+    "TradableProduct",
+    "Trade",
+    "TraderIdOrEmail",
+    "Unit",
+    "Unknown",
+    "UploadProductCatalogRequest",
+    "UploadProductCatalogResponse",
+    "UploadSymbologyRequest",
+    "UploadSymbologyResponse",
+    "UserId",
+    "Varying",
+    "Varying1",
+    "Venue",
+    "Withdrawal",
+    "WithdrawalsRequest",
+    "WithdrawalsResponse",
 ]