+++ examples/candles.py
@@ -1,9 +1,7 @@
 import asyncio
 
-from architect_py.async_client import AsyncClient
-from architect_py.common_types.tradable_product import TradableProduct
+from architect_py import AsyncClient, CandleWidth, TradableProduct
 from architect_py.graphql_client.exceptions import GraphQLClientHttpError
-from architect_py.grpc.models.definitions import CandleWidth
 
 from .common import connect_async_client